requirements-mcp.txt
mcp>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
gitpython>=3.1.0
anyio>=4.0.0
pytest>=7.0.0
